Clean up beacon committee subscriptions.
Beacon committee subscriptions were sending requests for all slots,
including slots that occurred in the past during startup.  This cleans
up the subscription request, only sending future slots at the time of
sending.
